backend/requirements.txt
Flask
flask-cors
flask-sqlalchemy
spacy
numpy
psycopg2-binary
beautifulsoup4
requests
Pillow
pytesseract
python-telegram-bot
moviepy
SpeechRecognition
m3u8
cloudscraper
twilio
blinker
selenium-wire 
opencv-python
gunicorn
flask_login
flask_caching
redis
flask_socketio
webdriver_manager
fake_useragent
SpeechRecognition

